# openapi.yml
openapi: 3.0.3
info:
  title: API Bancaria - Ejercicio Técnico
  version: 1.0.0
  description: API para gestionar Clientes, Cuentas, Movimientos y generacion de reportes (JSON / PDF base64).
servers:
  - url: /api
    description: API base (prefijo /api en paths)
...
paths:
  /clientes:
    get:
      summary: Listar clientes
      operationId: listClients
      tags:
        - Clients
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Lista de clientes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClienteDto'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
    post:
      summary: Crear cliente
      operationId: createClient
      tags:
        - Clients
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClienteCreateDto'
      responses:
        '201':
          description: Cliente creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClienteDto'
        '400':
          description: Validacion / datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
...
  /cuentas:
    get:
      summary: Listar cuentas
      operationId: listAccounts
      tags:
        - Accounts
      parameters:
        - name: clienteId
          in: query
          schema:
            type: string
            format: uuid
            description: Filtrar por cliente
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Lista de cuentas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CuentaDto'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
    post:
      summary: Crear cuenta
      operationId: createAccount
      tags:
        - Accounts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CuentaCreateDto'
      responses:
        '201':
          description: Cuenta creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CuentaDto'
        '400':
          description: Validacion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
...
  /movimientos:
    get:
      summary: Listar movimientos
      operationId: listMovements
      tags:
        - Movements
      parameters:
        - name: clienteId
          in: query
          schema:
            type: string
            format: uuid
        - name: numeroCuenta
          in: query
          schema:
            type: string
        - name: fechaDesde
          in: query
          schema:
            type: string
            format: date
        - name: fechaHasta
          in: query
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Lista de movimientos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MovimientoDto'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
    post:
      summary: Crear movimiento (débito/crédito)
      operationId: createMovement
      tags:
        - Movements
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovimientoCreateDto'
      responses:
        '201':
          description: Movimiento creado y saldo actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovimientoDto'
        '400':
          description: Error de validacion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '409':
          description: Conflicto de concurrencia (optimistic lock / version)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
...
components:
  schemas:
    PersonaDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        nombre:
          type: string
          maxLength: 200
        apellido:
          type: string
          maxLength: 200
        genero:
          type: string
          enum: ["MASCULINO", "FEMENINO", "OTRO"]
        fechaNacimiento:
          type: string
          format: date
        identificacion:
          type: string
          maxLength: 100
        telefono:
          type: string
          maxLength: 50
        direccion:
          type: string
        estado:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, nombre, estado]
      example:
        id: "9f2b7a2a-1d6f-4f1b-b2f0-123456789abc"
        nombre: "Jose"
        apellido: "Lema"
        genero: "MASCULINO"
        fechaNacimiento: "1985-01-15"
        identificacion: "0102030405"
        telefono: "098254785"
        direccion: "Otavalo sn y principal"
        estado: true
        createdAt: "2025-08-29T12:00:00Z"
        updatedAt: "2025-08-29T12:00:00Z"
    ClienteDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        personaId:
          type: string
          format: uuid
        persona:
          $ref: '#/components/schemas/PersonaDto'
        username:
          type: string
          maxLength: 100
        estado:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, personaId, username, estado]
      example:
        id: "2d1f3b9a-e4aa-4fa8-9f9c-abcdef012345"
        personaId: "9f2b7a2a-1d6f-4f1b-b2f0-123456789abc"
        persona:
          id: "9f2b7a2a-1d6f-4f1b-b2f0-123456789abc"
          nombre: "Jose"
          apellido: "Lema"
          genero: "MASCULINO"
          identificacion: "0102030405"
          telefono: "098254785"
          direccion: "Otavalo sn y principal"
          estado: true
        username: "jose.lema"
        estado: true
        createdAt: "2025-08-29T12:00:00Z"
        updatedAt: "2025-08-29T12:00:00Z"
    ClienteCreateDto:
      type: object
      properties:
        persona:
          $ref: '#/components/schemas/PersonaCreateDto'
        username:
          type: string
          maxLength: 100
        password:
          type: string
          format: password
          maxLength: 512
        estado:
          type: boolean
      required: [persona, username, password]
      example:
        persona:
          nombre: "Marianela"
          apellido: "Montalvo"
          genero: "FEMENINO"
          fechaNacimiento: "1990-03-15"
          identificacion: "1102233344"
          telefono: "098123456"
          direccion: "Av. Amazonas 123"
        username: "marianela.m"
        password: "Passw0rd!"
        estado: true
...
    PersonaCreateDto:
      type: object
      properties:
        nombre:
          type: string
          maxLength: 200
        apellido:
          type: string
          maxLength: 200
        genero:
          type: string
          enum: ["MASCULINO", "FEMENINO", "OTRO"]
        fechaNacimiento:
          type: string
          format: date
        identificacion:
          type: string
          maxLength: 100
        telefono:
          type: string
          maxLength: 50
        direccion:
          type: string
      required: [nombre, identificacion]
      example:
        nombre: "Juan"
        apellido: "Osorio"
        genero: "MASCULINO"
        fechaNacimiento: "1988-07-20"
        identificacion: "1203344556"
        telefono: "098874587"
        direccion: "Av. 6 de Diciembre 145"
...
    CuentaDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        clienteId:
          type: string
          format: uuid
        numeroCuenta:
          type: integer
          format: int64
        tipo:
          type: string
          enum: ["AHORROS", "CORRIENTE"]
        saldoInicial:
          type: number
          format: decimal
          multipleOf: 0.01
        saldoActual:
          type: number
          format: decimal
          multipleOf: 0.01
        estado:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, clienteId, numeroCuenta, tipo, saldoInicial, saldoActual, estado]
      example:
        id: "a1b2c3d4-e5f6-7890-1234-abcdef987654"
        clienteId: "2d1f3b9a-e4aa-4fa8-9f9c-abcdef012345"
        numeroCuenta: 478758
        tipo: "AHORROS"
        saldoInicial: 2000.00
        saldoActual: 1425.00
        estado: true
        createdAt: "2025-08-29T12:00:00Z"
        updatedAt: "2025-08-29T15:30:00Z"
    CuentaCreateDto:
      type: object
      properties:
        clienteId:
          type: string
          format: uuid
        tipo:
          type: string
          enum: ["AHORROS", "CORRIENTE"]
        saldoInicial:
          type: number
          format: decimal
          multipleOf: 0.01
          default: 0.00
        estado:
          type: boolean
          default: true
      required: [clienteId, tipo]
      example:
        clienteId: "2d1f3b9a-e4aa-4fa8-9f9c-abcdef012345"
        tipo: "CORRIENTE"
        saldoInicial: 1000.00
        estado: true
...
    MovimientoDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        cuentaId:
          type: string
          format: uuid
        fecha:
          type: string
          format: date-time
        tipo:
          type: string
          enum: ["DEBITO", "CREDITO"]
        valor:
          type: number
          format: decimal
          multipleOf: 0.01
        saldoPostMovimiento:
          type: number
          format: decimal
          multipleOf: 0.01
        createdAt:
          type: string
          format: date-time
      required: [id, cuentaId, fecha, tipo, valor, saldoPostMovimiento]
      example:
        id: "f1234567-89ab-4cde-0123-abcdefabcdef"
        cuentaId: "a1b2c3d4-e5f6-7890-1234-abcdef987654"
        fecha: "2025-08-28T10:30:00Z"
        tipo: "DEBITO"
        valor: 575.00
        saldoPostMovimiento: 1425.00
        createdAt: "2025-08-28T10:30:00Z"
    MovimientoCreateDto:
      type: object
      properties:
        cuentaId:
          type: string
          format: uuid
        tipo:
          type: string
          enum: ["DEBITO", "CREDITO"]
        valor:
          type: number
          format: decimal
          multipleOf: 0.01
      required: [cuentaId, tipo, valor]
      example:
        cuentaId: "a1b2c3d4-e5f6-7890-1234-abcdef987654"
        tipo: "DEBITO"
        valor: 575.00
...
    ErrorDto:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: string
      required: [timestamp, status, error, message]
      example:
        timestamp: "2025-08-29T12:10:00Z"
        status: 400
        error: "Bad Request"
        code: "SALDO_NO_DISPONIBLE"
        message: "Saldo no disponible para la operacion solicitada"
        details:
          - "Cuenta 478758 tiene saldo insuficiente"
          - "Saldo actual: 100.00, monto solicitado: 575.00"